mv scala test files to test sources folder.
