+++ src/test/scala/ValExec_MulAddRecFN.scala
@@ -35,8 +35,9 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 =============================================================================*/
 
-package hardfloat
+package hardfloat.test
 
+import hardfloat._
 import Chisel._
 
 class ValExec_MulAddRecFN(expWidth: Int, sigWidth: Int) extends Module
